start of work to add flux
Flux has a race condition where after submit (when we get
back a jobid) the file might still be needed. This means
that we need custom logic to not delete the temporary
file until cancel / worker completion. Flux also does
better getting an executable file, and a full path,
and so the submit function is modified for that.
Finally, flux does not support the concept of mem
or an account.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>